Move util functions to new file
