+++ angel_system/berkeley/utils/data/create_contact_metadata.py
@@ -15,6 +15,7 @@ import ubelt as ub
 import pickle
 import ast
 import math
+import pprint
 import random 
 
 from PIL import ImageFile
@@ -25,15 +26,12 @@ from demo import predictor, model
 from detectron2.utils.visualizer import Visualizer
 
 from utils.data.dataloaders.common import Re_order, time_from_name, preds_to_kwcoco
-
-#from dataloaders.load_kitware_coffee_data import coffee_activity_data_loader # Coffee specific
-from dataloaders.load_bbn_medical_data import bbn_activity_data_loader # M2 specific
-
+from utils.data.update_dets_utils import load_hl_hand_bboxes, replace_compound_label, find_closest_hands
 
 os.environ['CUDA_VISIBLE_DEVICES'] = '0'
 
 
-def load_model(config, conf_thr=0.4):
+def load_model(config, conf_thr=0.01):
     # Load model
     root_dir = '/home/local/KHQ/hannah.defazio/angel_system'
     berkeley_configs_dir = f"{root_dir}/angel_system/berkeley/configs"
@@ -89,16 +87,18 @@ def run_obj_detector(demo, bbn_root, data_dirs, split, no_contact=False, add_hl_
             predictions, step_infos, visualized_output = demo.run_on_image_smoothing_v2(
                     image, current_idx=idx)
             decoded_preds = model.decode_prediction(predictions)
+            using_contact = True if predictions[2] is not None else False
 
             if decoded_preds is not None:
                 preds[video_name][frame] = decoded_preds
                 
-                if no_contact:
-                    for class_ in preds[video_name][frame].keys():
-                        # Clear contact states
-                        preds[video_name][frame][class_]['obj_obj_contact_state'] = False
-                        preds[video_name][frame][class_]['obj_hand_contact_state'] = False
-                    
+                if no_contact and using_contact:
+                    for class_, dets in preds[video_name][frame].items():
+                        for i in range(len(dets)):
+                            # Clear contact states
+                            preds[video_name][frame][class_][i]['obj_obj_contact_state'] = False
+                            preds[video_name][frame][class_][i]['obj_hand_contact_state'] = False
+                        
                 # Image metadata needed later
                 preds[video_name][frame]['meta'] = {
                     'file_name': image_fn, #bbn_root + '/' + data_dir + '/' + video_name + '/_extracted/images/'  + image_fn.split('/')[-1],
@@ -120,159 +120,28 @@ def run_obj_detector(demo, bbn_root, data_dirs, split, no_contact=False, add_hl_
                             hand_bbox = [min(all_x_values), min(all_y_values), 
                                         max(all_x_values), max(all_y_values)] # tlbr
                             
-                            preds[video_name][frame][hand_label] = {
+                            new_det = {
                                 "confidence_score": 1,
                                 "bbox": hand_bbox,
-                                "obj_obj_contact_state": False,
-                                "obj_obj_contact_conf": 0,
-                                "obj_hand_contact_state": False,
-                                "obj_hand_contact_conf": 0,
                             }
+                            preds[video_name][frame][hand_label] = [new_det]
+
+                            if using_contact:
+                                con = {
+                                    "obj_obj_contact_state": False,
+                                    "obj_obj_contact_conf": 0,
+                                    "obj_hand_contact_state": False,
+                                    "obj_hand_contact_conf": 0,
+                                }
+
+                                preds[video_name][frame][hand_label] = [{**new_det, **con}]
 
             idx += 1
 
-    return preds
+    return preds, using_contact
 
-
-def load_hl_hand_bboxes(extracted_dir):
-    fn = extracted_dir + '/_hand_pose_2d_data.json'
-
-    if not os.path.exists(fn):
-        return {}
-    
-    with open(fn, 'r') as f:
-        hands = ast.literal_eval(f.read())
-    
-    all_hand_pose_2d = {}
-    for hand_info in hands:
-        time_stamp = float(hand_info["time_sec"]) + (float(hand_info["time_nanosec"]) * 1e-9)
-        if time_stamp not in all_hand_pose_2d.keys():
-            all_hand_pose_2d[time_stamp] = []
-
-        hand = hand_info["hand"].lower()
-        hand_label = f"hand ({hand})"
-
-        joints = {}
-        for joint in hand_info["joint_poses"]:
-            #if joint['clipped'] == 0:
-            joints[joint["joint"]] = joint # 2d position
-        if joints != {}:
-            all_hand_pose_2d[frame].append({
-                'hand': hand_label,
-                'joints': joints
-            })
-
-    return all_hand_pose_2d
-        
-
-def replace_compound_label(preds, obj, detected_classes, obj_hand_contact_state=False, obj_obj_contact_state=False):
-    """
-    Check if some subset of obj is in `detected_classes`
-
-    Designed to catch and correct incorrect compound classes
-    Ex: obj is "mug + filter cone + filter" but we detected "mug + filter cone"
-    """
-
-    compound_objs = [x.strip() for x in obj.split('+')]
-    detected_objs = [[y.strip() for y in x.split('+')]for x in detected_classes]
-
-    replaced_classes = []
-    for detected_class, objs in zip(detected_classes, detected_objs):
-        for obj_ in objs:
-            if obj_ in compound_objs:
-                replaced_classes.append(detected_class)
-                break
-
-    if replaced_classes == []:
-        # Case 0, We didn't detect any of the objects
-        replaced = None
-    elif len(replaced_classes) == 1:
-        # Case 1, we detected a subset of the compound as one detection
-        replaced = replaced_classes[0]
-        #print(f'replaced {replaced} with {obj}')
-
-        preds[obj] = preds.pop(replaced)
-        preds[obj]['obj_hand_contact_state'] = obj_hand_contact_state
-        preds[obj]['obj_obj_contact_state'] = obj_obj_contact_state
-    else:
-        # Case 2, the compound was detected as separate boxes
-        replaced = replaced_classes
-        #print(f'Combining {replaced} detections into compound \"{obj}\"')
-
-        new_bbox = None
-        new_conf = None
-        for det_obj in replaced:
-            bbox = preds[det_obj]['bbox']
-            conf = preds[det_obj]['confidence_score']
-
-            if new_bbox is None:
-                new_bbox = bbox
-            else:
-                # Find mix of bboxes
-                # TODO: first double check these are close enough that it makes sense to combine?
-                new_tlx, new_tly, new_brx, new_bry = new_bbox
-                tlx, tly, brx, bry = bbox
-
-                new_bbox = [min(new_tlx, tlx), min(new_tly, tly),
-                            max(new_brx, brx), max(new_bry, bry)]
-
-            new_conf = conf if new_conf is None else \
-                       (new_conf + conf) / 2 # average???
-
-            # remove old preds
-            preds.pop(det_obj)
-
-        new_pred = {
-            'confidence_score': new_conf,
-            'bbox': new_bbox,
-            'obj_obj_contact_state': obj_obj_contact_state,
-            'obj_hand_contact_state': obj_hand_contact_state
-        }
-        preds[obj] = new_pred
-    
-    return preds, replaced
-
-
-def find_closest_hands(object_pair, detected_classes, preds):
-    # Determine what the hand label is in the video, if any
-    # Fixes case where hand label has distinguishing information
-    # ex: hand(right) vs hand (left)
-
-    hand_labels = [h for h in detected_classes if 'hand' in h.lower()]
-    if len(hand_labels) == 0:
-        return None
-    
-    # find what object we should be interacting with
-    try:
-        obj = [o for o in object_pair if 'hand' not in o][0] # What to do if we don't have this???
-        obj_bbox = preds[obj]["bbox"]
-        w = abs(obj_bbox[2] - obj_bbox[0])
-        h = abs(obj_bbox[1] - obj_bbox[3])
-        obj_center = [obj_bbox[0] + (w/2), obj_bbox[1] + (h/2)]
-    except:
-        return None # TODO: temp???
-    
-    # Determine if any of the hands are close enough to the object to 
-    # likely be an interaction
-    min_dist = 180
-    close_hands = []
-    for i, hand_label in enumerate(hand_labels):
-        hand_bbox = preds[hand_label]["bbox"]
-        w = abs(hand_bbox[2] - hand_bbox[0])
-        h = abs(hand_bbox[1] - hand_bbox[3])
-        hand_center = [hand_bbox[0] + (w/2), hand_bbox[1] + (h/2)]
-        dist = math.dist(obj_center, hand_center)
-
-        if dist <= min_dist:
-            close_hands.append(hand_label)
-
-    hand_label = close_hands if len(close_hands) > 0 else None
-    return hand_label
-
-
-def update_preds(preds, step_map, data_root, data_dirs,
-                 no_contact=False, filter_activity_frames=True,
-                 using_step_labels=False, using_inter_steps=False, using_before_finished_task=False):
+def update_preds(activity_data_loader, preds, using_contact, original_step_map, step_map, data_root, data_dirs,
+                 experiment_flags):
     """
     Add the contact information back into the detections
     based on when the activities occurred and the contacts
@@ -289,11 +158,9 @@ def update_preds(preds, step_map, data_root, data_dirs,
 
         activity_only_preds[video_name] = {}
         
-        #gt_activity = coffee_activity_data_loader(video=video_name) # Coffee specific
-        gt_activity = bbn_activity_data_loader(videos_dir, video_name, step_map, lab_data,
-                                               using_inter_steps, using_before_finished_task)
-        #step_map['background'] = [['background', []]] # Add background to loop
-        #print('step map', step_map)
+        gt_activity = activity_data_loader(videos_dir, video_name, original_step_map, lab_data,
+                                           experiment_flags['using_inter_steps'], experiment_flags['using_before_finished_task'])
+
         for step, substeps in step_map.items():
             #print(step)
             inter_step = False
@@ -314,11 +181,14 @@ def update_preds(preds, step_map, data_root, data_dirs,
                                     if matching_gt['start'] <= ts <= matching_gt['end']}
                     matching_preds.update(matching_pred)
                 #print('matching preds', matching_preds)
+                #if matching_preds == {} and experiment_flags['filter_activity_frames']:
+                #    print('Skipping frame without activity ground truth')
+
                 for frame in matching_preds.keys():
                     detected_classes = list(preds[video_name][frame].keys())
                     preds[video_name][frame]["meta"]["activity_gt"] = f'{sub_step_str} ({step})' if not inter_step else sub_step_str
                     
-                    if not no_contact:
+                    if not experiment_flags['no_contact']:
                         # Keep original detections
                         continue
                     
@@ -350,19 +220,23 @@ def update_preds(preds, step_map, data_root, data_dirs,
                                 hand_labels = find_closest_hands(object_pair, detected_classes, preds[video_name][frame])
                                 if hand_labels is not None:
                                     found_items += 1
-                                    for hand_label in hand_labels:
-                                        preds[video_name][frame][hand_label]['obj_hand_contact_state'] = obj_hand_contact_state
-                                        
+                                    if using_contact:
+                                        for hand_label in hand_labels:
+                                            for i in range(len(preds[video_name][frame][hand_label])):
+                                                preds[video_name][frame][hand_label][i]['obj_hand_contact_state'] = obj_hand_contact_state
+                                            
                             elif obj in detected_classes:
                                 found_items += 1
-                                preds[video_name][frame][obj]['obj_hand_contact_state'] = obj_hand_contact_state
-                                preds[video_name][frame][obj]['obj_obj_contact_state'] = obj_obj_contact_state
-                                
+                                if using_contact:
+                                    for i in range(len(preds[video_name][frame][obj])):
+                                        preds[video_name][frame][obj][i]['obj_hand_contact_state'] = obj_hand_contact_state
+                                        preds[video_name][frame][obj][i]['obj_obj_contact_state'] = obj_obj_contact_state
+                                        
                             elif '+' in obj:
                                 # We might be missing part of a compound label
                                 # Let's try to fix that
                                 preds[video_name][frame], replaced = replace_compound_label(preds[video_name][frame], obj, detected_classes,
-                                                                  obj_hand_contact_state, obj_obj_contact_state)
+                                                                                            using_contact, obj_hand_contact_state, obj_obj_contact_state)
                                 detected_classes = list(preds[video_name][frame].keys())               
                                 if replaced is not None:
                                     found_items += 1
@@ -374,35 +248,49 @@ def update_preds(preds, step_map, data_root, data_dirs,
                             found.append(False)
 
                     # Add step number to class label
-                    if using_step_labels:
+                    if experiment_flags['using_step_labels']:
                         for class_ in detected_classes:
                             if class_ == 'meta':
                                 continue
-                            if '(step ' not in class_:
-                                print('updating label to include step')
+                            #print(class_)
+                            if '(step' not in class_ and '(before)' not in class_ and '(finished)' not in class_:
+                                #print('updating label to include step')
                                 preds[video_name][frame][f'{class_} ({step})'] = preds[video_name][frame].pop(class_)
 
                     detected_classes = list(preds[video_name][frame].keys())
             
-                    if filter_activity_frames:
+                    if experiment_flags['filter_all_obj_frames']:
                         if all(found):
                             print('Got all objects needed')
                             activity_only_preds[video_name][frame] = preds[video_name][frame]
                     else:
                         activity_only_preds[video_name][frame] = preds[video_name][frame]
 
-                    #else:
-                    #    print(step)
-                    #    print(detected_classes)
-    preds = activity_only_preds if no_contact else preds
+    preds = activity_only_preds if experiment_flags['filter_activity_frames'] else preds
     print("Updated contact metadata in predictions")
     return preds
 
-def main():
-    #demo = load_model(config='MC50-InstanceSegmentation/medical/M2/training_data/mask_rcnn_R_101_FPN_1x_BBN_M2_demo.yaml')
-    demo = load_model(config='MC50-InstanceSegmentation/medical/M2/preds/mask_rcnn_R_50_FPN_1x_BBN_M2_labels_with_inter_and_before_finished_steps_demo.yaml')
-    
-    metadata = demo.metadata.as_dict()
+
+def coffee_main():
+
+    coffee_root = '/Padlock_DT/Coffee'
+    ros_bags_dir = f'{coffee_root}/coffee_recordings/extracted'
+    videos_dir = ros_bags_dir # Coffee specific
+
+    training_split = {
+        'train': [f'all_activities_{x}' for x in [2, 10, 25, 28, 35]],
+        'val': [f'all_activities_{x}' for x in [24, 45]],
+        'test': [f'all_activities_{x}' for x in [41, 42, 43]],
+    } # Coffee specific
+
+    from dataloaders.load_kitware_coffee_data import coffee_activity_data_loader
+    activity_data_loader = coffee_activity_data_loader
+
+
+
+def tourniquet_main(stage, using_inter_steps, using_before_finished_task):
+    #demo = load_model(config='MC50-InstanceSegmentation/medical/M2/stage1/mask_rcnn_R_101_FPN_1x_BBN_M2_demo.yaml')
+    demo = load_model(config='MC50-InstanceSegmentation/medical/M2/stage2/mask_rcnn_R_50_FPN_1x_BBN_M2_labels_with_inter_and_before_finished_steps_demo.yaml')
 
     bbn_root = '/data/ptg/medical/bbn/data'
     data_root ='Release_v0.5/v0.52'
@@ -451,55 +339,13 @@ def main():
             'val': ['M2-72', 'M2-58', 'M2-134', 'M2-105', 'tq_18', 'M2-33', 'tq_22', 'M2-9', 'M2-42', 'M2-30', 'M2-16', 'M2-128'], 
             'test': ['tq_17', 'M2-119', 'M2-51', 'M2-31', 'M2-22', 'M2-36', 'M2-39', 'M2-70', 'tq_15', 'M2-4', 'M2-52', 'tq_7']
     }
-
     print(training_split)
 
-    """
-    coffee_root = '/Padlock_DT/Coffee'
-    ros_bags_dir = f'{coffee_root}/coffee_recordings/extracted'
-    videos_dir = ros_bags_dir # Coffee specific
+    from dataloaders.load_bbn_medical_data import bbn_activity_data_loader
+    activity_data_loader = bbn_activity_data_loader
 
-    training_split = {
-        'train': [f'all_activities_{x}' for x in [2, 10, 25, 28, 35]],
-        'val': [f'all_activities_{x}' for x in [24, 45]],
-        'test': [f'all_activities_{x}' for x in [41, 42, 43]],
-    } # Coffee specific
-    """
-    if not os.path.exists('temp'):
-        os.makedirs('temp')
-
-    experiment_name = 'm2_with_lab_cleaned_fixed_data_with_inter_and_before_finished_steps_no_contact'
-    stage = 'results'
-
-    print('Experiment: ', experiment_name)
-    print('Stage: ', stage)
-    
-    # Various experiment flags
-    if stage == 'results':
-        no_contact = False
-        filter_activity_frames = False
-    else:
-        no_contact = True
-        filter_activity_frames = True
-    
-    if no_contact:
-        print('Updating predictions with contact metadata')
-
-    if filter_activity_frames:
-        print("Filtering activity frames that do not include all relevant objects")
-    
-    using_step_labels = True
-    if using_step_labels:
-        print('Adding steps to labels')
-
-    using_inter_steps = True
-    if using_inter_steps:
-        print('Adding frames in-between activities')
-
-    using_before_finished_task = True
-    if using_before_finished_task:
-        print('Adding frames from before and after the activities')
-
+    # Update step map
+    metadata = demo.metadata.as_dict()
 
     step_map = metadata['sub_steps']
     if using_inter_steps:
@@ -510,24 +356,49 @@ def main():
     if using_before_finished_task:
         step_map['before'] = [['Not started'.lower(), [['tourniquet_tourniquet', 'hand']]]]
         step_map['finished'] = [['Finished'.lower(), [['tourniquet_tourniquet', 'hand']]]]
-    print('step map: ', step_map)
+    print(f'step map: {step_map}')
+    
+    return demo, training_split, bbn_root, data_dirs, activity_data_loader, metadata, step_map
+
+
+def main():
+    experiment_name = 'm2_with_lab_cleaned_fixed_data_with_inter_and_before_finished_steps_no_contact'
+    stage = 'results'
+
+    print('Experiment: ', experiment_name)
+    print('Stage: ', stage)
+    
+    # Various experiment flags
+    experiment_flags = {
+        'no_contact': False if stage == 'results' else True,
+        'filter_activity_frames': False if stage == 'results' else True,
+        'filter_all_obj_frames': False,
+        'using_step_labels' : True,
+        'using_inter_steps': True,
+        'using_before_finished_task': True
+    }
+    print(f'experiment flags: {experiment_flags}')
+
+    demo, training_split, data_root, data_dirs, activity_data_loader, metadata, step_map = tourniquet_main(stage, experiment_flags['using_inter_steps'], experiment_flags['using_before_finished_task'])
     
     if stage == 'stage2':
         splits = ['val', 'train_contact', 'test', 'train_activity']
     else:
-        splits = ['val', 'test', 'train_activity']
+        splits = ['test', 'train_activity', 'val']
 
     #splits = ['test']
     #data_dirs = (kitware_test_dir, lab_data_dir)
     #training_split['test'] = kitware_videos
     for split in splits:
         print(f'{split}: {len(training_split[split])} videos')
+
         # Raw detector output
-        preds_no_contact = run_obj_detector(demo, 
-                                            bbn_root, data_dirs,
+        preds_no_contact, using_contact = run_obj_detector(demo, 
+                                            data_root, data_dirs,
                                             training_split[split], 
-                                            no_contact=no_contact, 
+                                            no_contact=experiment_flags['no_contact'], 
                                             add_hl_hands=False)
+        print(f'Using contact: {using_contact}')
         
         fn = f'temp/{experiment_name}_{split}_preds_no_contact.pickle'
         print('temp file: ', fn)
@@ -536,18 +407,19 @@ def main():
             pickle.dump(preds_no_contact, fh)
 
         # Update contact info based on gt
-        preds_with_contact = update_preds(preds_no_contact,
+        preds_with_contact = update_preds(activity_data_loader,
+                                          preds_no_contact,
+                                          using_contact,
+                                          metadata['sub_steps'],
                                           step_map, 
-                                          bbn_root, data_dirs, 
-                                          no_contact,
-                                          filter_activity_frames,
-                                          using_step_labels, using_inter_steps, using_before_finished_task)
+                                          data_root, data_dirs, 
+                                          experiment_flags)
         
         dset = preds_to_kwcoco(metadata, preds_with_contact, '', save_fn=f'{experiment_name}_{stage}_{split}.mscoco.json',
                                # assuming detector already has the right labels so these aren't needed here
-                               using_step_labels=False if stage == 'results' else using_step_labels,
-                               using_inter_steps=False if stage == 'results' else using_inter_steps,
-                               using_before_finished_task=False if stage == 'results' else using_before_finished_task)
+                               using_step_labels=False if stage == 'results' else experiment_flags['using_step_labels'],
+                               using_inter_steps=False if stage == 'results' else experiment_flags['using_inter_steps'],
+                               using_before_finished_task=False if stage == 'results' else experiment_flags['using_before_finished_task'])
 
     # TODO: train on save_fn + save model
